Rename client typings to client.d.ts
